+++ nuget-client/BrevgeneratorClient.Tests.ArchUnit/Layers.cs
@@ -1,0 +1,81 @@
+using ArchUnitNET.Domain;
+using static ArchUnitNET.Fluent.ArchRuleDefinition;
+
+namespace Arbeidstilsynet.Brevgenerator.Client.Test.ArchUnit
+{
+    internal static class Constants
+    {
+        internal static string NameSpacePrefix = @"Arbeidstilsynet\.Brevgenerator\.Client";
+        internal static string OwnedBrevgeneratorClientFullName =
+            "Arbeidstilsynet.Brevgenerator.Client.DependencyInjection.OwnedBrevgeneratorClient";
+
+        internal static string CoverageCollectorNamespace = "Microsoft.CodeCoverage.Instrumentation.Static.Tracker";
+        internal static string RootNamespace = $"^({NameSpacePrefix}|{NameSpacePrefix}\\..*)$";
+        internal static string ExtensionsNamespace = CreateNamespaceRegex("Extensions");
+        internal static string DependencyInjectionNamespace = CreateNamespaceRegex("DependencyInjection");
+        internal static string ModelsNamespace = CreateNamespaceRegex("Models");
+
+        private static string CreateNamespaceRegex(string namespaceSection)
+        {
+            return $@"^({NameSpacePrefix}\.{namespaceSection}|{NameSpacePrefix}\.{namespaceSection}\..*|{NameSpacePrefix}\..*\.{namespaceSection}|{NameSpacePrefix}\..*\.{namespaceSection}\..*)$";
+        }
+    }
+
+    internal static class Layers
+    {
+        internal static readonly System.Reflection.Assembly BrevgeneratorClientAssembly =
+            typeof(Arbeidstilsynet.Brevgenerator.Client.IAssemblyInfo).Assembly;
+
+        internal static readonly System.Reflection.Assembly SystemConsoleAssembly = typeof(System.Console).Assembly;
+
+        internal static readonly IObjectProvider<IType> BrevgeneratorClientLayer = Types()
+            .That()
+            .ResideInAssembly(BrevgeneratorClientAssembly)
+            .And()
+            .DoNotResideInNamespace(Constants.CoverageCollectorNamespace)
+            .As("BrevgeneratorClient Layer");
+
+        internal static readonly IObjectProvider<IType> PublicInterfaces = Interfaces()
+            .That()
+            .Are(BrevgeneratorClientLayer)
+            .And()
+            .ArePublic()
+            .As("public interfaces");
+
+        internal static readonly IObjectProvider<IType> InterfaceImplementations = Classes()
+            .That()
+            .Are(BrevgeneratorClientLayer)
+            .And()
+            .AreAssignableTo(PublicInterfaces)
+            .And()
+            .AreNot(PublicInterfaces)
+            .And()
+            .DoNotHaveFullName(Constants.OwnedBrevgeneratorClientFullName)
+            .As("interface implementations");
+
+        internal static readonly IObjectProvider<IType> PublicAbstractClasses = Classes()
+            .That()
+            .Are(BrevgeneratorClientLayer)
+            .And()
+            .AreAbstract()
+            .And()
+            .ArePublic()
+            .As("public abstract classes");
+
+        internal static readonly IObjectProvider<IType> ExportableTypes = Types()
+            .That()
+            .ResideInNamespaceMatching(Constants.ExtensionsNamespace)
+            .Or()
+            .ResideInNamespaceMatching(Constants.DependencyInjectionNamespace)
+            .Or()
+            .ResideInNamespaceMatching(Constants.ModelsNamespace)
+            .As("inside exportable namespaces");
+
+        internal static readonly IObjectProvider<IType> TypesInInternalNamespaces = Types()
+            .That()
+            .Are(BrevgeneratorClientLayer)
+            .And()
+            .AreNot(ExportableTypes)
+            .As("outside exportable namespaces");
+    }
+}