+++ nuget-client/BrevgeneratorClient.Tests/IAssemblyInfo.cs
@@ -1,0 +1,3 @@
+namespace Arbeidstilsynet.Brevgenerator.Client.Test;
+
+public interface IAssemblyInfo { }